Insert halaman Pemain

+++ app/Http/Livewire/Admin/ClubListIndex.php
@@ -4,6 +4,7 @@ namespace App\Http\Livewire\Admin;
 
 use Livewire\Component;
 use App\ClubList;
+use App\Player;
 use Livewire\WithPagination;
 use Illuminate\Support\Facades\Session;
 
@@ -13,6 +14,8 @@ class ClubListIndex extends Component
     public $countData;
     public $paginate = 10;
     public $search;
+
+    protected $listeners = ['delete', 'add_player'];
 
     public function mount()
     {
@@ -37,12 +40,22 @@ class ClubListIndex extends Component
     public function delete($id)
     {
         if($id){
-            $data = ClubList::find($id);
-            $data->delete();    
-            //flash message
-            session()->flash('message', array('type' => 'success', 'content' => 'Data berhasil dihapus!'));
+            $cek_pemain = Player::where('club_list_id', $id)->first();
+            if(empty($cek_pemain)){
+                $data = ClubList::find($id);
+                $data->delete();    
+                //flash message
+                session()->flash('message', array('type' => 'success', 'content' => 'Data berhasil dihapus!'));
+            }else{
+                session()->flash('message', array('type' => 'error', 'content' => 'Data gagal dihapus, karena ada pemainnya!'));
+            }
             //redirect
             return redirect()->route('admin.list-club');
         }
+    }
+
+    public function add_player($id)
+    {
+        return redirect()->route('admin.player-create', ['id' => $id]);
     }
 }
